+++ src/app/components/sidebar-spaces/space-item.js
@@ -8,18 +8,23 @@
  */
 
 import { computed, get } from '@ember/object';
-import { reads } from '@ember/object/computed';
+import { reads, collect } from '@ember/object/computed';
 import { inject as service } from '@ember/service';
 import Component from '@ember/component';
 import _ from 'lodash';
 import bytesToString from 'onedata-gui-common/utils/bytes-to-string';
 import computedPipe from 'onedata-gui-common/utils/ember/computed-pipe';
 import I18n from 'onedata-gui-common/mixins/components/i18n';
+import safeExec from 'onedata-gui-common/utils/safe-method-execution';
 
 export default Component.extend(I18n, {
   tagName: '',
 
   i18n: service(),
+  globalNotify: service(),
+  spaceActions: service(),
+  router: service(),
+  navigationState: service(),
 
   /**
    * @override
@@ -43,6 +48,16 @@ export default Component.extend(I18n, {
    * @type {boolean}
    */
   inSidenav: false,
+
+  /**
+   * @type {boolean}
+   */
+  leaveSpaceModalOpen: false,
+
+  /**
+   * @type {boolean}
+   */
+  isLeaving: false,
 
   /**
    * Just an one-way alias
@@ -107,4 +122,58 @@ export default Component.extend(I18n, {
    * @type {Ember.ComputedProperty<string>}
    */
   _totalSupportSizeHumanReadable: computedPipe('_totalSupportSize', bytesToString),
+
+  /**
+   * @type {Ember.ComputedProperty<Utils.Action>}
+   */
+  leaveAction: computed(function leaveAction() {
+    return {
+      action: () => this.send('showLeaveModal'),
+      title: this.t('leave'),
+      class: 'leave-record-trigger',
+      icon: 'leave-space',
+    };
+  }),
+
+  /**
+   * @type {Ember.ComputedProperty<Utils.Action>}
+   */
+  removeAction: computed('space', function removeAction() {
+    const {
+      spaceActions,
+      space,
+    } = this.getProperties('spaceActions', 'space');
+
+    return spaceActions.createRemoveSpaceAction({ space });
+  }),
+
+  /**
+   * @type {ComputedProperty<Array<Utils.Action>>}
+   */
+  itemActions: collect('leaveAction', 'removeAction'),
+
+  actions: {
+    showLeaveModal() {
+      this.set('leaveSpaceModalOpen', true);
+    },
+    closeLeaveModal() {
+      this.set('leaveSpaceModalOpen', false);
+    },
+    leave() {
+      const {
+        space,
+        spaceActions,
+        navigationState,
+      } = this.getProperties('space', 'spaceActions', 'navigationState');
+      this.set('isLeaving', true);
+      return spaceActions.leaveSpace(space)
+        .then(() => navigationState.redirectToCollectionIfResourceNotExist())
+        .finally(() =>
+          safeExec(this, 'setProperties', {
+            isLeaving: false,
+            leaveSpaceModalOpen: false,
+          })
+        );
+    },
+  },
 });