+++ src/app/locales/en/utils/space-actions/remove-space-action.js
@@ -1,0 +1,10 @@
+export default {
+  title: 'Remove',
+  modalHeader: 'Remove space',
+  modalDescription: 'You are about to delete the space <strong>{{spaceName}}</strong>.',
+  modalDisclaimer: '<strong>This operation is irreversible</strong>. All the data in the space will be lost, although for some storage backends it might be possible to import the storage data into another space.',
+  modalCheckboxMessage: 'I understand and want to proceed.',
+  modalYes: 'Remove',
+  successNotificationText: 'The space has been sucessfully removed.',
+  failureNotificationActionName: 'removing the space',
+};